ontology: separate link path selection from tokens
